PDFTASK-1 Add a new replace method and a constant to ParsedPdfMapperImpl, Add @RequiredArgsConstructor to PdfController, Add GlobalExceptionHandler, Fixed try-catch logic in PdfParser, removed duplicate code from PdfServiceImpl
