--- src/main/java/com/example/pdfboxandshit/service/PdfService.java
@@ -1,13 +1,0 @@
-package com.example.pdfboxandshit.service;
-
-import com.example.pdfboxandshit.dto.ParsePdfDto;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-
-public interface PdfService {
-
-    String getFullText(MultipartFile file) throws IOException;
-
-    ParsePdfDto parsePdf(MultipartFile file) throws IOException;
-}
