+++ src/main/java/com/example/pdfparsingtask/exception/PdfParsingException.java
@@ -1,0 +1,8 @@
+package com.example.pdfparsingtask.exception;
+
+public class PdfParsingException extends RuntimeException {
+    public PdfParsingException(String message, Throwable cause) {
+
+        super(message, cause);
+    }
+}
