+++ src/main/java/com/example/pdfparsingtask/model/ParsedPdfModel.java
@@ -1,33 +1,21 @@
-package com.example.pdfboxandshit.repository.entity;
+package com.example.pdfparsingtask.model;
 
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
-import java.util.UUID;
 
 @Data
 @Builder
-@Table(name = "documents")
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-public class ParsedPdfEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
+public class ParsedPdfModel {
 
     private Long fileSize;
+
     private String fileExtension;
 
     private String lastName;
     private String firstName;
     private String patronymic;
-
     private LocalDate birthDate;
     private String taxpayerIdentificationNumber;
     private Integer documentType;
