+++ src/main/java/com/example/pdfparsingtask/repository/ParsePdfRepository.java
@@ -1,6 +1,6 @@
-package com.example.pdfboxandshit.repository;
+package com.example.pdfparsingtask.repository;
 
-import com.example.pdfboxandshit.repository.entity.ParsedPdfEntity;
+import com.example.pdfparsingtask.repository.entity.ParsedPdfEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
