+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=PDFBoxAndShit
+spring.application.name=PDFParsingTask
 spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
 spring.datasource.username=adming
 spring.datasource.password=admin
