+++ src/test/java/com/example/pdfparsingtask/PdfParsingTaskApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.pdfboxandshit;
+package com.example.pdfparsingtask;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PdfBoxAndShitApplicationTests {
+class PdfParsingTaskApplicationTests {
 
     @Test
     void contextLoads() {
